weight init trait
